Add a GUI
This is a simple Tk gui that shows the current config values. It also
shows connection state and does some config sanity checking. It needs
only a reconnnect button and some thread-state management for the 1st
launch user-case

It defaults to enabled on Windows but not on linux or when Tk is missing

Written in Tk only for no extra dependencies and a simpler single-file
executable experience.

A bit of thread/async management mess here because of asyncio, but seems
to work well.